Lots of changes:
(1) gcc and so on does not live in /lib/, changed configure to contain a more reasonable location.  Also, gcc-4 is non-standard and I changed that too.
(2) Wrote a tool to calculate the angular distortion between two objects.
(3) Updated CAT_LBConformalMap.cpp to take mapScale and the location of the P point on the command line.


git-svn-id: svn://141.35.69.214/CAT@4 85a05d3d-4204-4b9d-9be8-300d97d0f21f